Split out TempSocket, introduce modules
TempSocket is primarily a space-saving optimization: it takes 32 bytes
and a duplicate heap allocation off of Socket, and is slightly more
ergonomic to use.

This also moves SshMux into its own module, and TempSocket into a module
on ssh_mux.

I also added some doc comments on the two structs, and subsumed the
changes from #18 and #11.
